+++ srcs/entities/move.c
@@ -6,7 +6,7 @@
 /*   By: nlecaill <nlecaill@student.le-101.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/18 10:39:59 by dgascon           #+#    #+#             */
-/*   Updated: 2020/03/13 10:44:47 by nlecaill         ###   ########lyon.fr   */
+/*   Updated: 2020/03/13 11:39:46 by nlecaill         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 int		move_x(t_data *data, float value)
 {
-	float tmp;
-	char block;
-	int	lenx;
+	float	tmp;
+	char	block;
+	int		lenx;
 
 	tmp = data->player.pos.x + value;
-	lenx = data->world.leny[data->player.pos.y / BLOCK_SIZE];
+	lenx = data->world.leny[(int)(data->player.pos.y / BLOCK_SIZE)];
 	if (tmp < (lenx * BLOCK_SIZE) && tmp > BLOCK_SIZE)
 	{
-		block = data->world.map[data->player.pos.y / BLOCK_SIZE]
+		block = data->world.map[(int)(data->player.pos.y / BLOCK_SIZE)]
 					[(int)(tmp / BLOCK_SIZE)];
 		if (!(block >= '1' && block <= '9'))
 			data->player.pos.x = tmp;
@@ -33,13 +33,13 @@ int		move_x(t_data *data, float value)
 int		move_y(t_data *data, float value)
 {
 	float	tmp;
-	char		block;
+	char	block;
 
 	tmp = data->player.pos.y + value;
 	if (tmp < (data->world.size.y * BLOCK_SIZE) && tmp > BLOCK_SIZE)
 	{
 		block = data->world.map[(int)(tmp / BLOCK_SIZE)]
-					[data->player.pos.x / BLOCK_SIZE];
+					[(int)(data->player.pos.x / BLOCK_SIZE)];
 		if (!(block >= '1' && block <= '9'))
 			data->player.pos.y = tmp;
 	}
